convert invite user to modal
