+++ deepfence_frontend/apps/dashboard/src/routes/private.tsx
@@ -72,7 +72,6 @@ import { module as editGlobalSettings } from '@/features/settings/pages/EditGlob
 import { module as editUser } from '@/features/settings/pages/EditUser';
 import { module as emailConfiguration } from '@/features/settings/pages/EmailConfiguration';
 import { module as globalSettings } from '@/features/settings/pages/GlobalSettings';
-import { module as inviteUser } from '@/features/settings/pages/InviteUser';
 import { module as settings } from '@/features/settings/pages/Settings';
 import { module as userAuditLogs } from '@/features/settings/pages/UserAuditLogs';
 import { module as userManagement } from '@/features/settings/pages/UserManagement';
@@ -497,11 +496,6 @@ export const privateRoutes: CustomRouteObject[] = [
             path: 'user-management/edit/:userId',
             ...editUser,
             meta: { title: 'Edit User Account' },
-          },
-          {
-            path: 'user-management/invite-user',
-            ...inviteUser,
-            meta: { title: 'Invite User' },
           },
           {
             path: 'user-audit-logs',
